refactor: avoid import InlineDiffService directly
