+++ packages/ai-native/src/browser/widget/inline-stream-diff/live-preview.component.tsx
@@ -46,40 +46,6 @@ export interface IRemovedWidgetState {
   position: IPosition;
 }
 
-export enum EPartialEdit {
-  accept = 'accept',
-  discard = 'discard',
-}
-
-export interface IPartialEditEvent {
-  uri: URI;
-  /**
-   * 总 diff 数
-   */
-  totalPartialEditCount: number;
-  /**
-   * 已处理的个数
-   */
-  resolvedPartialEditCount: number;
-  /**
-   * 已添加行数
-   */
-  totalAddedLinesCount: number;
-  /**
-   * 已采纳的个数
-   */
-  acceptPartialEditCount: number;
-  /**
-   * 已删除行数
-   */
-  totalDeletedLinesCount: number;
-  currentPartialEdit: {
-    type: EPartialEdit;
-    addedLinesCount: number;
-    deletedLinesCount: number;
-  };
-}
-
 export interface ITextLinesTokens {
   text: string;
   lineTokens: LineTokens;
